requirements.txt
BioPython
pandas
tqdm
numpy
scipy
pyensembl
openpyxl
logomaker
seaborn
# tRNAscan-SE  bioconda
# blast bioconda